Move grdcontour to grdcontour.py
