+++ pygmt/src/__init__.py
@@ -5,6 +5,7 @@ Source code for PyGMT modules.
 # GMT Primary modules
 from pygmt.src.coast import coast
 from pygmt.src.colorbar import colorbar
+from pygmt.src.grdcontour import grdcontour
 from pygmt.src.logo import logo
 
 # GMT Supplementary modules
